// Assets-Uncapped/front-portal-develop/src/services/api/organisation-users/models/ProspectWithEligibilityCheckResponse.ts
export interface ProspectWithEligibilityCheckResponse {
    /**
     * 
     * @type {boolean}
     * @memberof ProspectWithEligibilityCheckResponse
     */
    eligibleForFunding?: boolean;
    /**
     * 
     * @type {number}
     * @memberof ProspectWithEligibilityCheckResponse
     */
    preliminaryOffer?: number;
    /**
     * 
     * @type {string}
     * @memberof ProspectWithEligibilityCheckResponse
     */
    preliminaryOfferCurrency?: string;
    /**
     * 
     * @type {string}
     * @memberof ProspectWithEligibilityCheckResponse
     */
    notEligibleReason?: ProspectWithEligibilityCheckResponseNotEligibleReasonEnum;
    /**
     * 
     * @type {string}
     * @memberof ProspectWithEligibilityCheckResponse
     */
    prospectId?: string;
    /**
     * 
     * @type {number}
     * @memberof ProspectWithEligibilityCheckResponse
     */
    minimumRequiredRevenue?: number;
    /**
     * 
     * @type {string}
     * @memberof ProspectWithEligibilityCheckResponse
     */
    minimumRequiredRevenueCurrency?: string;
}


/**
 * @export
 */
export const ProspectWithEligibilityCheckResponseNotEligibleReasonEnum = {
    SoleTrader: 'SOLE_TRADER',
    UnsupportedUsState: 'UNSUPPORTED_US_STATE',
    UnsupportedCanadaProvince: 'UNSUPPORTED_CANADA_PROVINCE',
    UnsupportedBusinessType: 'UNSUPPORTED_BUSINESS_TYPE',
    UnsupportedCountry: 'UNSUPPORTED_COUNTRY',
    LowRevenue: 'LOW_REVENUE'
} as const;
export type ProspectWithEligibilityCheckResponseNotEligibleReasonEnum = typeof ProspectWithEligibilityCheckResponseNotEligibleReasonEnum[keyof typeof ProspectWithEligibilityCheckResponseNotEligibleReasonEnum];